requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
fake-useragent==0.1.11
idna==2.10
Pillow==7.2.0
pytesseract==0.3.5
requests==2.24.0
soupsieve==2.0.1
urllib3==1.25.10
